Restructure code placement
