+++ packages/Pagination/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "dist",
     "declarationDir": "dist/types"
   },
-  "include": ["src"]
+  "include": ["src", "examples"]
 }
